requirements.txt
Flask
flask-restx
python-dotenv
marshmallow
pytest
pytest-mock
pytest-cov
Flask-Migrate
psycopg2-binary
Flask-SQLAlchemy
Faker
Flask-HTTPAuth
Werkzeug
SQLAlchemy
alembic
flask-login
marshmallow-sqlalchemy
elasticsearch